BRAIN-TUMOUR-DETECTION/bt/requirements.txt
-f https://download.pytorch.org/whl/cpu/torch_stable.html
Flask_Cors==3.0.10
numpy==1.22.0
dill==0.3.4
opencv_python_headless==4.5.5.62
Flask==2.0.2
Pillow==9.0.0
torch==1.10.1+cpu
torchvision==0.11.2+cpu
gunicorn==20.1.0